All tests are successful on `FreeBSD'.
git-svn-id: svn://svn.savannah.gnu.org/myserver/trunk@2828 c4c922ea-a70d-0410-937d-ec1daf91936c

+++ myserver/tests/test_regex.cpp
@@ -55,8 +55,8 @@ public:
     CPPUNIT_ASSERT_EQUAL(ret, 0);
 
 
-    CPPUNIT_ASSERT_EQUAL(pm[1].rm_so, 6);
-    CPPUNIT_ASSERT_EQUAL(pm[1].rm_eo, 11);
+    CPPUNIT_ASSERT_EQUAL((int)pm[1].rm_so, 6);
+    CPPUNIT_ASSERT_EQUAL((int)pm[1].rm_eo, 11);
   }
 
   void testStarMatch()
